batch-layer/requirements.txt
#Hầu hết thư viện được cài đặt từ khi setup tổng quan ban đầu rồi, đây không cần thiết phải cài lại
# Batch Layer Requirements
# Kafka
kafka-python==2.0.2

# HDFS
hdfs==2.7.0

# Data Processing
pandas==2.1.3
pyarrow==14.0.1

# Configuration
pyyaml==6.0.1

# Monitoring
prometheus-client==0.19.0

# Cassandra (for Phase 4 Step 4)
cassandra-driver==3.29.1

# PySpark (for Phase 4 Step 3)
pyspark==3.5.0
